+++ app/lib/frontend/templates/views/pkg/package_list.dart
@@ -6,7 +6,6 @@ import 'package:_pub_shared/search/search_form.dart';
 import 'package:_pub_shared/search/tags.dart';
 import 'package:clock/clock.dart';
 import 'package:pana/pana.dart';
-import 'package:pub_dev/frontend/request_context.dart';
 
 import '../../../../package/models.dart';
 import '../../../../package/screenshots/backend.dart';
@@ -252,7 +251,7 @@ d.Node _item({
               classes: ['packages-description'],
               children: [
                 d.span(text: description),
-                if (requestContext.experimentalFlags.showTopics) ...topics,
+                ...topics,
               ],
             ),
             d.p(classes: ['packages-metadata'], child: metadataNode),